move client radiation methods to own class
